backend/app/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
python-dotenv
psycopg2-binary